refactor: create credentials at start app
