+++ apps/vs-agent/src/controllers/public/self-tr/index.ts
@@ -1,2 +1,1 @@
 export * from './SelfTrController'
-export * from './SelfTrService'
