+++ apps/vs-agent/src/public.module.ts
@@ -1,6 +1,6 @@
 import { DynamicModule, Module } from '@nestjs/common'
 
-import { DidWebController, InvitationRoutesController, SelfTrController, SelfTrService } from './controllers'
+import { DidWebController, InvitationRoutesController, SelfTrController } from './controllers'
 import { VsAgentService } from './services/VsAgentService'
 import { VsAgent } from './utils/VsAgent'
 
@@ -20,7 +20,6 @@ export class PublicModule {
           provide: 'PUBLIC_API_BASE_URL',
           useValue: publicApiBaseUrl,
         },
-        SelfTrService,
         VsAgentService,
       ],
       exports: [],
